+++ lib/ssl_pinning_plugin.dart
@@ -2,11 +2,11 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 
-
-enum SHA {SHA1, SHA256}
+enum SHA { SHA1, SHA256 }
 
 class SslPinningPlugin {
-  static const MethodChannel _channel = const MethodChannel('ssl_pinning_plugin');
+  static const MethodChannel _channel =
+      const MethodChannel('ssl_pinning_plugin');
 
   // Création d'un singleton pour le plugin
   static final SslPinningPlugin _sslPinning = new SslPinningPlugin._internal();
@@ -17,13 +17,18 @@ class SslPinningPlugin {
     _channel.setMethodCallHandler(_platformCallHandler);
   }
 
-  static Future<String> check({ String serverURL, Map<String, String> headerHttp, SHA sha, List<String> allowedSHAFingerprints, int timeout }) async {
+  static Future<String> check(
+      {String serverURL,
+      Map<String, String> headerHttp,
+      SHA sha,
+      List<String> allowedSHAFingerprints,
+      int timeout}) async {
     final Map<String, dynamic> params = <String, dynamic>{
-      "url" : serverURL,
-      "headers" : headerHttp,
+      "url": serverURL,
+      "headers": headerHttp,
       "type": sha.toString().split(".").last,
-      "fingerprints" : allowedSHAFingerprints,
-      "timeout" : timeout
+      "fingerprints": allowedSHAFingerprints,
+      "timeout": timeout
     };
     String resp = await _channel.invokeMethod('check', params);
     return resp;